chore: created a component for the Form
